backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-chess==1.999
httpx==0.28.1
pydantic==2.11.7
pytest==8.4.2
